requirements.txt
discord-webhook~=1.4.1
beautifulsoup4~=4.13.3
lxml~=5.3.1
beautifulsoup4~=4.13.3
pyopenmensa~=0.95.0
PyYAML~=6.0.2